changed chit number to string format

+++ src/main/java/com/eminence/chitty/jwt/entity/Manager.java
@@ -39,13 +39,13 @@ public class Manager {
     private String passWord;
 
     @Column(name = "roleid")
-    private Integer roleId;
+    private Long roleId;
 
     @JsonIgnore
     @OneToMany(cascade = CascadeType.ALL,mappedBy = "manager",fetch = FetchType.LAZY)
     private Set<Chitty> chits= new HashSet<>();
 
-    public Manager(Long emp_id, String firstName, String emp_lastname, String email, Long mobileNumber, String passWord, Integer roleId, Set<Chitty> chits) {
+    public Manager(Long emp_id, String firstName, String emp_lastname, String email, Long mobileNumber, String passWord, Long roleId, Set<Chitty> chits) {
         this.emp_id = emp_id;
         this.firstName = firstName;
         this.emp_lastname = emp_lastname;
@@ -108,11 +108,11 @@ public class Manager {
         this.passWord = passWord;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
